✨ (common): Implement NewPayload Function in Common
